pyproject.toml
[tool.poetry]
name = "cli-benchmark"
version = "0.1.0"
description = ""
authors = ["Hans Bricks <33793809+gyrusdentatus@users.noreply.github.com>>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic = "^2.7.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
